requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.6
qdrant-client>=1.11.0
openai>=1.40.0
anthropic>=0.34.0
python-docx>=1.1.2
PyPDF2>=3.0.1
openpyxl>=3.1.4
tiktoken>=0.7.0
google-auth>=2.30.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.131.0
python-dotenv>=1.0.1
pydantic>=2.9.0
httpx>=0.27.0
python-multipart>=0.0.9
rich>=13.7.1
typer>=0.12.3
pyyaml>=6.0
discord.py>=2.3.0